+++ backend/src/main/java/ca/metricalsky/winston/entity/Video.java
@@ -26,7 +26,7 @@ public class Video {
 
     private String thumbnailUrl;
 
-    private Boolean commentsDisabled;
+    private boolean commentsDisabled;
 
     private OffsetDateTime publishedAt;