+++ backend/src/main/java/ca/metricalsky/winston/entity/fetch/FetchAction.java
@@ -64,6 +64,7 @@ public class FetchAction {
     }
 
     public enum Status {
+        READY,
         FETCHING,
         COMPLETED,
         FAILED,