+++ backend/src/main/java/ca/metricalsky/winston/service/NotificationsService.java
@@ -1,15 +1,14 @@
 package ca.metricalsky.winston.service;
 
+import ca.metricalsky.winston.config.exception.AppProblemDetail;
 import ca.metricalsky.winston.events.SubscriptionEvent;
 import ca.metricalsky.winston.exception.AppException;
-import ca.metricalsky.winston.utils.SsePublisher;
+import ca.metricalsky.winston.events.SsePublisher;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.SmartLifecycle;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
-import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
-import java.io.IOException;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
@@ -17,19 +16,21 @@ import java.util.concurrent.ConcurrentHashMap;
 
 @Slf4j
 @Service
-public class NotificationsService {
+public class NotificationsService implements SmartLifecycle {
 
     private static final Long DEFAULT_TIMEOUT_MS = 30_000L;
 
-    private final Map<UUID, SseEmitter> subscriptions = new ConcurrentHashMap<>();
+    private final Map<UUID, SsePublisher> subscriptions = new ConcurrentHashMap<>();
+    private boolean running;
 
-    public SsePublisher openSubscription() throws IOException {
+    public SsePublisher openSubscription() {
         return openSubscription(DEFAULT_TIMEOUT_MS);
     }
 
-    public SsePublisher openSubscription(Long timeout) throws IOException {
-        var subscriptionId = UUID.randomUUID();
-        var sseEmitter = new SseEmitter(timeout);
+    public SsePublisher openSubscription(Long timeout) {
+        var ssePublisher = new SsePublisher(timeout);
+        var subscriptionId = ssePublisher.getId();
+        var sseEmitter = ssePublisher.getSseEmitter();
         sseEmitter.onCompletion(() -> {
             log.info("Notifications subscription {} completed", subscriptionId);
             subscriptions.remove(subscriptionId);
@@ -42,17 +43,40 @@ public class NotificationsService {
             log.error("Notifications subscription {} failed", subscriptionId, exception);
             subscriptions.remove(subscriptionId);
         });
-        subscriptions.put(subscriptionId, sseEmitter);
+        subscriptions.put(subscriptionId, ssePublisher);
 
         log.info("Notifications subscription {} opened", subscriptionId);
-        sseEmitter.send(new SubscriptionEvent(true, subscriptionId), MediaType.APPLICATION_JSON);
-        return new SsePublisher(sseEmitter);
+        ssePublisher.publish(new SubscriptionEvent(true, subscriptionId));
+        return ssePublisher;
     }
 
     public SsePublisher requireSubscription(UUID subscriptionId) {
         return Optional.ofNullable(subscriptions.get(subscriptionId))
-                .map(SsePublisher::new)
                 .orElseThrow(() -> new AppException(HttpStatus.UNPROCESSABLE_ENTITY,
                         "The provided subscription stream is not open."));
+    }
+
+    @Override
+    public void start() {
+        running = true;
+    }
+
+    @Override
+    public void stop() {
+        if (!subscriptions.isEmpty()) {
+            var ex = new AppException(HttpStatus.SERVICE_UNAVAILABLE,
+                    "The server is shutting down and is no longer available.");
+            Map.copyOf(subscriptions).forEach((subscriptionId, ssePublisher) -> {
+                log.warn("Closing notifications subscription {}", subscriptionId);
+                ssePublisher.publish(AppProblemDetail.forException(ex));
+                ssePublisher.completeWithError(ex);
+            });
+        }
+        running = false;
+    }
+
+    @Override
+    public boolean isRunning() {
+        return running;
     }
 }